test-task-pets-master/requirements.txt
dj-database-url==0.4.2
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.6.4
gunicorn==19.7.1
MarkupSafe==1.0
psycopg2==2.7.4
pytz==2017.2
whitenoise==3.3.1
